// src/types/deal.ts
export type DealDetails = {
  id: string;
  name: string;
  address: string;
  city: string;
  state: string;
  propertyType: string;
  dateUpdated: string;
  imageUrl: string;
  seller: string;
  guidancePrice: number;
  guidancePricePSF: number;
  capRate: number;
  propertySize: number;
  landArea: number;
  zoning: string;
  pdfUrl: string;
};